backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
celery
redis
faiss-cpu
numpy
openai
pymupdf
python-docx
python-multipart
pyyaml
